rpc improvements in groups. restructure code. fix jetbrains run config.
