+++ agent/agent.go
@@ -6,11 +6,9 @@ package agent
 
 import (
 	"crypto/tls"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"github.com/eclipse/paho.mqtt.golang"
-	"github.com/ns1labs/orb"
 	"github.com/ns1labs/orb/agent/backend"
 	"github.com/ns1labs/orb/fleet"
 	"go.uber.org/zap"
@@ -82,38 +80,6 @@ func (a *orbAgent) nameAgentRPCTopics() {
 	a.heartbeatsTopic = fmt.Sprintf("%s/%s", base, fleet.HeartbeatsTopic)
 	a.logTopic = fmt.Sprintf("%s/%s", base, fleet.LogTopic)
 
-}
-func (a *orbAgent) sendSingleHeartbeat(t time.Time, state fleet.State) {
-
-	a.logger.Debug("heartbeat")
-
-	hbData := fleet.Heartbeat{
-		SchemaVersion: fleet.CurrentHeartbeatSchemaVersion,
-		TimeStamp:     t,
-		State:         state,
-	}
-
-	body, err := json.Marshal(hbData)
-	if err != nil {
-		a.logger.Error("error creating heartbeat data", zap.Error(err))
-		return
-	}
-
-	if token := a.client.Publish(a.heartbeatsTopic, 1, false, body); token.Wait() && token.Error() != nil {
-		a.logger.Error("error sending heartbeat", zap.Error(token.Error()))
-	}
-}
-
-func (a *orbAgent) sendHeartbeats() {
-	a.sendSingleHeartbeat(time.Now(), fleet.Online)
-	for {
-		select {
-		case <-a.hbDone:
-			return
-		case t := <-a.hbTicker.C:
-			a.sendSingleHeartbeat(t, fleet.Online)
-		}
-	}
 }
 
 func (a *orbAgent) unsubscribeGroupChannels() {
@@ -208,89 +174,11 @@ func (a *orbAgent) Start() error {
 	return nil
 }
 
-func (a *orbAgent) sendCapabilities() error {
-
-	capabilities := fleet.Capabilities{
-		SchemaVersion: fleet.CurrentCapabilitiesSchemaVersion,
-		AgentTags:     a.config.OrbAgent.Tags,
-		OrbAgent: fleet.OrbAgentInfo{
-			Version: orb.GetVersion(),
-		},
-	}
-
-	capabilities.Backends = make(map[string]fleet.BackendInfo)
-	for name, be := range a.backends {
-		ver, err := be.Version()
-		if err != nil {
-			a.logger.Error("backend failed to retrieve version", zap.String("backend", name), zap.Error(err))
-			continue
-		}
-		capabilities.Backends[name] = fleet.BackendInfo{
-			Version: ver,
-		}
-	}
-
-	body, err := json.Marshal(capabilities)
-	if err != nil {
-		return err
-	}
-
-	if token := a.client.Publish(a.capabilitiesTopic, 1, false, body); token.Wait() && token.Error() != nil {
-		return token.Error()
-	}
-
-	return nil
-}
-
-func (a *orbAgent) sendGroupMembershipReq() error {
-
-	payload := fleet.GroupMembershipReqRPCPayload{}
-
-	data := fleet.RPC{
-		SchemaVersion: fleet.CurrentRPCSchemaVersion,
-		Func:          fleet.GroupMembershipReqRPCFunc,
-		Payload:       payload,
-	}
-
-	body, err := json.Marshal(data)
-	if err != nil {
-		return err
-	}
-
-	if token := a.client.Publish(a.rpcToCoreTopic, 1, false, body); token.Wait() && token.Error() != nil {
-		return token.Error()
-	}
-
-	return nil
-}
-
-func (a *orbAgent) sendAgentPoliciesReq() error {
-
-	payload := fleet.AgentPoliciesReqRPCPayload{}
-
-	data := fleet.RPC{
-		SchemaVersion: fleet.CurrentRPCSchemaVersion,
-		Func:          fleet.AgentPoliciesReqRPCFunc,
-		Payload:       payload,
-	}
-
-	body, err := json.Marshal(data)
-	if err != nil {
-		return err
-	}
-
-	if token := a.client.Publish(a.rpcToCoreTopic, 1, false, body); token.Wait() && token.Error() != nil {
-		return token.Error()
-	}
-
-	return nil
-}
-
-func (a *orbAgent) subscribeGroupChannels(channels []string) []string {
+func (a *orbAgent) subscribeGroupChannels(groups []fleet.GroupMembershipData) []string {
 	var successList []string
-	for _, channelID := range channels {
+	for _, groupData := range groups {
 
-		base := fmt.Sprintf("channels/%s/messages", channelID)
+		base := fmt.Sprintf("channels/%s/messages", groupData.ChannelID)
 		rpcFromCoreTopic := fmt.Sprintf("%s/%s", base, fleet.RPCFromCoreTopic)
 
 		token := a.client.Subscribe(rpcFromCoreTopic, 1, a.handleGroupRPCFromCore)
@@ -307,106 +195,10 @@ func (a *orbAgent) subscribeGroupChannels(channels []string) []string {
 			a.logger.Error("failed to subscribe to group channel/topic: time out", zap.String("topic", rpcFromCoreTopic))
 			continue
 		}
-		a.logger.Info("subscribed to a group", zap.String("topic", rpcFromCoreTopic))
-		successList = append(successList, channelID)
+		a.logger.Info("completed RPC subscription to group", zap.String("name", groupData.Name), zap.String("topic", rpcFromCoreTopic))
+		successList = append(successList, groupData.ChannelID)
 	}
 	return successList
-}
-
-func (a *orbAgent) handleGroupMembership(rpc fleet.GroupMembershipRPCPayload) {
-
-	// if this is the full list, reset all group subscriptions and subscribed to this list
-	if rpc.FullList {
-		a.unsubscribeGroupChannels()
-		a.groupChannels = a.subscribeGroupChannels(rpc.ChannelIDS)
-	} else {
-		// otherwise, just add these subscriptions to the existing list
-		successList := a.subscribeGroupChannels(rpc.ChannelIDS)
-		a.groupChannels = append(a.groupChannels, successList...)
-	}
-}
-
-func (a *orbAgent) handleAgentPolicies(rpc []fleet.AgentPolicyRPCPayload) {
-
-	a.logger.Info("Agent Policy from core", zap.Any("payload", rpc))
-
-}
-
-func (a *orbAgent) handleGroupRPCFromCore(client mqtt.Client, message mqtt.Message) {
-
-	a.logger.Info("Group RPC message from core", zap.String("topic", message.Topic()), zap.ByteString("payload", message.Payload()))
-
-	var rpc fleet.RPC
-	if err := json.Unmarshal(message.Payload(), &rpc); err != nil {
-		a.logger.Error("error decoding RPC message from core", zap.Error(fleet.ErrSchemaMalformed))
-		return
-	}
-	if rpc.SchemaVersion != fleet.CurrentRPCSchemaVersion {
-		a.logger.Error("error decoding RPC message from core", zap.Error(fleet.ErrSchemaVersion))
-		return
-	}
-	if rpc.Func == "" || rpc.Payload == nil {
-		a.logger.Error("error decoding RPC message from core", zap.Error(fleet.ErrSchemaMalformed))
-		return
-	}
-
-	// dispatch
-	switch rpc.Func {
-	case fleet.AgentPolicyRPCFunc:
-		var r fleet.AgentPolicyRPC
-		if err := json.Unmarshal(message.Payload(), &r); err != nil {
-			a.logger.Error("error decoding agent policy message from core", zap.Error(fleet.ErrSchemaMalformed))
-			return
-		}
-		a.handleAgentPolicies(r.Payload)
-	default:
-		a.logger.Warn("unsupported/unhandled core RPC, ignoring",
-			zap.String("func", rpc.Func),
-			zap.Any("payload", rpc.Payload))
-	}
-
-}
-
-func (a *orbAgent) handleRPCFromCore(client mqtt.Client, message mqtt.Message) {
-
-	a.logger.Info("RPC message from core", zap.String("topic", message.Topic()), zap.ByteString("payload", message.Payload()))
-
-	var rpc fleet.RPC
-	if err := json.Unmarshal(message.Payload(), &rpc); err != nil {
-		a.logger.Error("error decoding RPC message from core", zap.Error(fleet.ErrSchemaMalformed))
-		return
-	}
-	if rpc.SchemaVersion != fleet.CurrentRPCSchemaVersion {
-		a.logger.Error("error decoding RPC message from core", zap.Error(fleet.ErrSchemaVersion))
-		return
-	}
-	if rpc.Func == "" || rpc.Payload == nil {
-		a.logger.Error("error decoding RPC message from core", zap.Error(fleet.ErrSchemaMalformed))
-		return
-	}
-
-	// dispatch
-	switch rpc.Func {
-	case fleet.GroupMembershipRPCFunc:
-		var r fleet.GroupMembershipRPC
-		if err := json.Unmarshal(message.Payload(), &r); err != nil {
-			a.logger.Error("error decoding group membership message from core", zap.Error(fleet.ErrSchemaMalformed))
-			return
-		}
-		a.handleGroupMembership(r.Payload)
-	case fleet.AgentPolicyRPCFunc:
-		var r fleet.AgentPolicyRPC
-		if err := json.Unmarshal(message.Payload(), &r); err != nil {
-			a.logger.Error("error decoding agent policy message from core", zap.Error(fleet.ErrSchemaMalformed))
-			return
-		}
-		a.handleAgentPolicies(r.Payload)
-	default:
-		a.logger.Warn("unsupported/unhandled core RPC, ignoring",
-			zap.String("func", rpc.Func),
-			zap.Any("payload", rpc.Payload))
-	}
-
 }
 
 func (a *orbAgent) Stop() {
